+++ Zenith/Models/Company.cs
@@ -49,7 +49,7 @@ namespace Zenith.Models
         public string Address { get; set; } = string.Empty;
 
         [Reactive]
-        public long CreditValue { get; set; }
+        public float CreditValue { get; set; }
 
         [Required(AllowEmptyStrings = true)]
         [MaxLength(LengthConstants.VERY_LARGE_STRING)]
